Fix modal exit
On a previous version, I was selecting a div overlay, but it wasn't the
correct one. There are a few of them in the document, and the first one
wasn't the one used by Ayuda. So we clicked on an inactive overlay.

This version even manages to make it generic for the Ayuda sub-modals
(like Lámina) and even for confirmation modals like Finalizar examen.